Complete refactor of classes
Instead of using separate js files linked to the html file, the
classes are imported and exported. Improves the naming consistency,
as now classes are named Interface instead of UI.
The button is captured in the main script file instead of the
constructor of the Game class, so the global scope of the script is now
responsible for the starting of the game, making the code more readable.